refactor: move around come CLI code
