+++ stac-cli/src/lib.rs
@@ -1,9 +1,7 @@
 mod args;
+mod command;
 mod error;
 
-pub use {
-    args::{Args, Command},
-    error::Error,
-};
+pub use {args::Args, command::Command, error::Error};
 
 pub type Result<T> = std::result::Result<T, Error>;
